backend/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools",
]

[project]
name = "msm"
description = "MAAS Site Manager"
readme = "README.md"
maintainers = [
  { name = "MAAS Developers", email = "maas-devel@lists.launchpad.net" },
]
authors = [
  { name = "MAAS Developers", email = "maas-devel@lists.launchpad.net" },
]
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
]
dynamic = [
  "version",
]
dependencies = [
  "alembic[tz]",
  "baize",
  "bcrypt==4.0.1",
  "beautifulsoup4",
  "boto3",
  "fastapi==0.115.9",                                  # manually pinning because of https://github.com/encode/starlette/issues/2893 -- introduced by Starlette version 46.0, which is pinned by FastAPI>=0.115.10
  "jinja2",
  "passlib[bcrypt]",                                   # manually pinning because of https://foss.heptapod.net/python-libs/passlib/-/issues/190 FIXME: can be removed once passlib is updated to 1.7.5
  "prometheus-fastapi-instrumentator",
  "pydantic[email]",
  "pydantic-settings",
  "python-jose[cryptography]",
  "python-multipart",
  "pytz",
  "sqlalchemy[postgresql-asyncpg,postgresql-psycopg]",
  "streaming-form-data",
  "strenum",                                           # emulate enum.StrEnum, which is available in Python3.11
  "temporal-lib-py>=1.8",
  "ujson",                                             # faster JSON encoding/decoding (for fastapi)
  "uvicorn[standard]",
  "uvloop",
]
optional-dependencies.testing = [
  "httpx",
  "psycopg",
  "pytest",
  "pytest-asyncio",
  "pytest-mock",
  "pytest-postgresql",
  "pytest-xdist",
]
scripts.msm-admin = "msm.cmd.admin:script"
scripts.msm-api = "msm.api:run"
scripts.msm-sampledata = "msm.cmd.sampledata:script"

[tool.setuptools.package-data]
msm = [
  "db/alembic/alembic.ini",
  "db/alembic/script.py.mako",
]

[tool.setuptools.dynamic]
version = { attr = "msm.__version__" }

[tool.setuptools.packages.find]
include = [
  "msm*",
  "temporal",
]

[tool.ruff]
line-length = 79

lint.select = [
  "F401", # unused imports
  "I",    # isort
  "RUF",  # ruff
  "UP",   # pyupgrade
]
lint.isort.combine-as-imports = true
lint.isort.force-sort-within-sections = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = [
  "tests/",
]
junit_logging = "all"
junit_suite_name = "maas-site-manager backend tests"

[tool.coverage.report]
fail_under = 80.0
show_missing = true
skip_covered = true

[tool.coverage.run]
source = [
  "msm",
  "temporal",
]
omit = [
  "msm/db/alembic/env.py",
  "msm/sampledata/*",
  "msm/cmd/sampledata/*",
]

[tool.mypy]
install_types = true
junit_xml = "../junit-mypy.xml"
non_interactive = true
plugins = [
  "pydantic.mypy",
]
strict = true
